+++ config/query-builder.php
@@ -36,8 +36,11 @@ return [
     'disable_invalid_filter_query_exception' => false,
 
     /*
-     * By default the package inspects query string of request - $request->query().
-     * You can change this behavior to $request->input() by setting this value to `true`.
+     * By default the package inspects query string of request using $request->query().
+     * You can change this behavior to inspect the request body using $request->input()
+     * by setting this value to `body`.
+     *
+     * Possible values: `query_string`, `body`
      */
-    'inspect_input_instead_of_query' => false,
+    'request_data_source' => 'query_string',
 ];